+++ angular/projects/wcs-formly/src/lib/formly-wcs-field-radio.component.ts
@@ -5,10 +5,7 @@ import { FieldType } from '@ngx-formly/core';
   // tslint:disable-next-line:component-selector
   selector: 'app-formly-wcs-field-radio',
   template: `
-    <wcs-form-field [attr.is-error]="showError ? true : null">
-      <wcs-label *ngIf="to.label && to.hideLabel !== true" [attr.for]="id">
-        {{ to.label }}
-      </wcs-label>
+    <formly-wcs-field-wrapper [field]="field" [id]="id" [showError]="showError" [to]="to">
       <wcs-radio-group
         [attr.mode]="to?.attributes?.mode"
         [attr.required]="to.required && to.hideRequiredMarker !== true"
@@ -21,11 +18,7 @@ import { FieldType } from '@ngx-formly/core';
           [value]="option.value"
           [label]="option.label"></wcs-radio>
       </wcs-radio-group>
-      <wcs-error *ngIf="showError">
-        <formly-validation-message #error [field]="field"></formly-validation-message>
-      </wcs-error>
-      <wcs-hint *ngIf="to.description">{{ to.description }}</wcs-hint>
-    </wcs-form-field>
+    </formly-wcs-field-wrapper>
   `,
   styles: []
 })