+++ angular/projects/wcs-formly/src/lib/formly-wcs-field-select.component.ts
@@ -5,10 +5,7 @@ import { FieldType } from '@ngx-formly/core';
   // tslint:disable-next-line:component-selector
   selector: 'app-formly-wcs-field-select',
   template: `
-    <wcs-form-field [attr.is-error]="showError ? true : null">
-      <wcs-label *ngIf="to.label && to.hideLabel !== true" [attr.for]="id">
-        {{ to.label }}
-      </wcs-label>
+    <formly-wcs-field-wrapper [field]="field" [id]="id" [showError]="showError" [to]="to">
       <wcs-select
         [id]="to.id"
         [attr.required]="(to.required && to.hideRequiredMarker !== true) ? true : null"
@@ -21,11 +18,7 @@ import { FieldType } from '@ngx-formly/core';
           [attr.disabled]="option.disabled ? true : null"
           [value]="option.value">{{option.label}}</wcs-select-option>
       </wcs-select>
-      <wcs-error *ngIf="showError">
-        <formly-validation-message #error [field]="field"></formly-validation-message>
-      </wcs-error>
-      <wcs-hint *ngIf="to.description">{{ to.description }}</wcs-hint>
-    </wcs-form-field>
+    </formly-wcs-field-wrapper>
   `,
   styles: []
 })