+++ angular/projects/wcs-formly/src/public-api.ts
@@ -8,4 +8,6 @@ export * from './lib/formly-wcs-field-select.component';
 export * from './lib/formly-wcs-field-textarea.component';
 export * from './lib/formly-wcs-field-checkbox.component';
 export * from './lib/formly-wcs-field-switch.component';
+export * from './lib/formly-wcs-field-wrapper.component';
+export * from './lib/formly-wcs-boolean-field-wrapper.component';
 export * from './lib/wcs-formly.module';